% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_input}
\alias{check_input}
\title{Check variable inputs}
\usage{
check_input(
  input,
  input_type,
  input_validation = NULL,
  min_value = -Inf,
  max_value = Inf,
  max_length = 1L,
  variable_name = NULL,
  is_missing = NULL
)
}
\arguments{
\item{input}{.}

\item{input_type}{.}

\item{input_validation}{.}

\item{min_value}{.}

\item{max_value}{.}

\item{max_length}{.}

\item{variable_name}{.}

\item{is_missing}{.}
}
\value{
nothing
}
\description{
Internal function to check input variables
}
