% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account_status.R
\name{get_account_status}
\alias{get_account_status}
\title{Get the status of your GTmetrix account}
\usage{
get_account_status(api_key)
}
\arguments{
\item{api_key}{An active GTmetrix API key. (string)}
}
\value{
A data.frame that contains meta data of a GTmetrix account.
}
\description{
Show available credits and other meta data for the suppliad API key.
}
\examples{
\dontrun{output_table <- get_account_status(
                           api_key = "API_KEY"
                         )}
}
