% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_tests.R
\name{get_all_tests}
\alias{get_all_tests}
\title{get_all_tests}
\usage{
get_all_tests(api_key, page_size = 50, page_number = 1)
}
\arguments{
\item{api_key}{An active GTmetrix API key. (string)}

\item{page_size}{Page size (default 50, max 500)}

\item{page_number}{Page (default 1)}
}
\value{
A data.frame object that contains test IDs and their meta data.
}
\description{
Get a table of tests, their report IDs and other meta data.
}
\examples{
\dontrun{output_table <- get_all_tests(api_key = "API_KEY")}
}
