% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_browser_details.R
\name{get_browser_details}
\alias{get_browser_details}
\title{Show browser details}
\usage{
get_browser_details(browser_id, api_key)
}
\arguments{
\item{browser_id}{Browser ID. (integer)}

\item{api_key}{An active GTmetrix API key. (string)}
}
\value{
A data.frame object that contains available browsers and their meta data.
}
\description{
Get details for a specific browsers ID.
}
\examples{
\dontrun{output_table <- get_browser_details(browser_id = 3, api_key = "API_KEY")}
}
