% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_location_details.R
\name{get_location_details}
\alias{get_location_details}
\title{Show location details}
\usage{
get_location_details(location, api_key)
}
\arguments{
\item{location}{Location ID. (integer)}

\item{api_key}{An active GTmetrix API key. (string)}
}
\value{
A data.frame object that contains available locations
and their meta data.
}
\description{
Get details for a specific locations ID.
}
\examples{
\dontrun{output_table <- get_location_details(
                             location_id = 3, api_key = "API_KEY"
                         )}
}
