% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report.R
\name{get_report}
\alias{get_report}
\title{Get status and meta data of a specific report}
\usage{
get_report(report_id, api_key)
}
\arguments{
\item{report_id}{ID of a GTmetrix report. (string)}

\item{api_key}{An active GTmetrix API key. (string)}
}
\value{
A data.frame object that contains a GTmetrix report and its
meta data.
}
\description{
Get status and meta data of a specific GTmetrix report.
}
\examples{
\dontrun{output_table <- get_report(
                           test_id = "REPORT_ID",
                           api_key = "API_KEY"
                         )}
}
