% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_available_browsers.R
\name{show_available_browsers}
\alias{show_available_browsers}
\title{Show available browsers}
\usage{
show_available_browsers(api_key)
}
\arguments{
\item{api_key}{An active GTmetrix API key. (string)}
}
\value{
A data.frame object that contains available browsers and their meta data.
}
\description{
Show available browsers for the supplied API key.
}
\examples{
\dontrun{output_table <- show_available_browsers(api_key = "API_KEY")}
}
