% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_available_locations.R
\name{show_available_locations}
\alias{show_available_locations}
\title{Show available locations}
\usage{
show_available_locations(api_key)
}
\arguments{
\item{api_key}{An active GTmetrix API key. (string)}
}
\value{
A data.frame object that contains available locations and their
meta data.
}
\description{
Show available locations for the supplied API key.
}
\examples{
\dontrun{output_table <- show_available_locations(api_key = "API_KEY")}
}
