% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_test.R
\name{start_test}
\alias{start_test}
\title{start_test}
\usage{
start_test(url, api_key, wait_for_completion = TRUE, ...)
}
\arguments{
\item{url}{The web url to test. (https:// or http:// are optional)}

\item{api_key}{An active GTmetrix API key}

\item{wait_for_completion}{Whether the function should wait for the completion of the test. If TRUE (default), the report associated with the test ID will be requested in roughly 3 second intervals and returned, when successful. If FALSE, the meta data of the test will be returned.}
}
\value{
A data frame object that contains either the test meta data or the GTmetrix report.
}
\description{
Starts a GTmetrix test and returns the test or the report
}
\examples{
\dontrun{output_table <- start_test("google.com", api_key = "API_KEY", wait_for_completion = TRUE)}
}
